requirements.txt
# Happy Hour Discovery System Requirements
# GPT-5 enabled OpenAI client and dependencies

# Core OpenAI GPT-5 support
openai>=1.40.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Environment and utilities
python-dotenv>=1.0.0

# Optional: For enhanced data processing
numpy>=1.24.0
python-dateutil>=2.8.2

# Optional: For progress bars in terminal
tqdm>=4.65.0

# Optional: For better JSON handling
orjson>=3.9.0